--- Lista2/README.md
@@ -1,1 +1,0 @@
-# NEED TO WRITE
